feat: 인증 토큰 헤더 추가

+++ app/src/main/java/woowacourse/shopping/data/api/ProductApi.kt
@@ -18,6 +18,6 @@ interface ProductApi {
 
     @GET("/products/{id}")
     fun getProductDetail(
-        @Path("id") productId: Int,
+        @Path("id") productId: Long,
     ): Call<ProductDetailResponse>
 }
