+++ app/src/main/java/woowacourse/shopping/data/mapper/CartMapper.kt
@@ -18,7 +18,7 @@ fun CartItemsResponse.Content.toDomain(): CartProduct =
         id = id,
         productDetail =
             ProductDetail(
-                id = product.id.toInt(),
+                id = product.id,
                 name = product.name,
                 imageUrl = product.imageUrl,
                 price = product.price,
