+++ app/src/main/java/woowacourse/shopping/data/model/entity/CartProductEntity.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "cart_products")
 data class CartProductEntity(
-    @PrimaryKey val productId: Int,
+    @PrimaryKey val productId: Long,
     val quantity: Int,
 )
