+++ app/src/main/java/woowacourse/shopping/data/model/entity/HistoryProductEntity.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "search_history")
 data class HistoryProductEntity(
-    @PrimaryKey val productId: Int,
+    @PrimaryKey val productId: Long,
     val timestamp: Long = System.currentTimeMillis(),
 )
