+++ app/src/main/java/woowacourse/shopping/ui/catalog/adapter/history/HistoryProductViewHolder.kt
@@ -17,6 +17,6 @@ class HistoryProductViewHolder(
     }
 
     fun interface OnClickHandler {
-        fun onHistoryProductClick(id: Int)
+        fun onHistoryProductClick(id: Long)
     }
 }
